@@ -1,10 +1,10 @@
 buildscript {
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.0'
+        classpath 'com.android.tools.build:gradle:3.3.0'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -13,8 +13,8 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
         google()
+        jcenter()
     }
 }
 
@@ -27,24 +27,30 @@ ext {
     // Sdk and tools
     // Support library and architecture components support minSdk 14 and above.
     minSdkVersion = 14
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '26.0.2'
+    targetSdkVersion = 28
+    compileSdkVersion = 28
 
     // App dependencies
-    supportLibraryVersion = '26.1.0'
-    guavaVersion = '22.0-android'
+    androidXVersion = '1.0.0'
+    androidXTestVersion = '1.1.1'
+    androidXAnnotations = '1.0.1'
+    androidXLegacySupport = '1.0.0'
+    materialVersion = '1.0.0'
+    cardVersion = '1.0.0'
+    recyclerViewVersion = '1.0.0'
+    appCompatVersion = '1.0.2'
+
+    guavaVersion = '27.0.1-android'
     junitVersion = '4.12'
     mockitoVersion = '1.10.19'
-    powerMockito = '1.6.2'
+    dexMaker = '1.2'
     hamcrestVersion = '1.3'
     runnerVersion = '1.0.1'
     rulesVersion = '1.0.1'
-    espressoVersion = '3.0.1'
-
-    // Architecture Components dependencies
-    roomVersion = "1.0.0"
-    archLifecycleVersion = "1.0.0-rc1"
+    espressoVersion = '3.1.1'
+    roomVersion = '2.0.0'
+    archLifecycleVersion = '2.0.0'
+    archTestingVersion = '1.1.1'
 }
 
 /*
