@@ -16,19 +16,13 @@
 
 package com.example.android.architecture.blueprints.todoapp.statistics;
 
+import android.app.Application;
 import android.content.Intent;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.IdlingRegistry;
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
 
-import com.example.android.architecture.blueprints.todoapp.Injection;
 import com.example.android.architecture.blueprints.todoapp.R;
 import com.example.android.architecture.blueprints.todoapp.ViewModelFactory;
 import com.example.android.architecture.blueprints.todoapp.data.Task;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository;
-import com.example.android.architecture.blueprints.todoapp.taskdetail.TaskDetailActivity;
 import com.example.android.architecture.blueprints.todoapp.util.EspressoIdlingResource;
 
 import org.junit.After;
@@ -37,10 +31,16 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import androidx.test.core.app.ApplicationProvider;
+import androidx.test.espresso.IdlingRegistry;
+import androidx.test.filters.LargeTest;
+import androidx.test.rule.ActivityTestRule;
+import androidx.test.runner.AndroidJUnit4;
+
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.containsString;
 
 /**
@@ -59,30 +59,27 @@ public class StatisticsScreenTest {
      */
     @Rule
     public ActivityTestRule<StatisticsActivity> mStatisticsActivityTestRule =
-            new ActivityTestRule<>(StatisticsActivity.class, true, false);
+            new ActivityTestRule<>(StatisticsActivity.class, false, false);
 
-    /**
-     * Setup your test fixture with a fake task id. The {@link TaskDetailActivity} is started with
-     * a particular task id, which is then loaded from the service API.
-     *
-     * <p>
-     * Note that this test runs hermetically and is fully isolated using a fake implementation of
-     * the service API. This is a great way to make your tests more reliable and faster at the same
-     * time, since they are isolated from any outside dependencies.
-     */
     @Before
-    public void intentWithStubbedTaskId() {
-        // Given some tasks
-        ViewModelFactory.destroyInstance();
-        TasksRepository tasksRepository = Injection.provideTasksRepository(
-                InstrumentationRegistry.getTargetContext());
-        tasksRepository.deleteAllTasks();
-        tasksRepository.saveTask(new Task("Title1", "", false));
-        tasksRepository.saveTask(new Task("Title2", "", true));
+    public void prepareRepository() throws Throwable {
+        mStatisticsActivityTestRule.runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                TasksRepository repo = ViewModelFactory.getInstance(
+                        (Application) ApplicationProvider.getApplicationContext())
+                        .getTasksRepository();
 
-        // Lazily start the Activity from the ActivityTestRule
-        Intent startIntent = new Intent();
-        mStatisticsActivityTestRule.launchActivity(startIntent);
+                repo.deleteAllTasks();
+                repo.saveTask(new Task("St1", "", false));
+                repo.saveTask(new Task("St2", "", true));
+            }
+        });
+
+        Intent intent = new Intent(ApplicationProvider.getApplicationContext(),
+                StatisticsActivity.class);
+
+        mStatisticsActivityTestRule.launchActivity(intent);
     }
 
     /**
@@ -96,22 +93,22 @@ public class StatisticsScreenTest {
         IdlingRegistry.getInstance().register(EspressoIdlingResource.getIdlingResource());
     }
 
-    @Test
-    public void Tasks_ShowsNonEmptyMessage() throws Exception {
-        // Check that the active and completed tasks text is displayed
-        String expectedActiveTaskText = InstrumentationRegistry.getTargetContext()
-                .getString(R.string.statistics_active_tasks, 1);
-        onView(withText(containsString(expectedActiveTaskText))).check(matches(isDisplayed()));
-        String expectedCompletedTaskText = InstrumentationRegistry.getTargetContext()
-                .getString(R.string.statistics_completed_tasks, 1);
-        onView(withText(containsString(expectedCompletedTaskText))).check(matches(isDisplayed()));
-    }
-
     /**
      * Unregister your Idling Resource so it can be garbage collected and does not leak any memory.
      */
     @After
     public void unregisterIdlingResource() {
         IdlingRegistry.getInstance().unregister(EspressoIdlingResource.getIdlingResource());
+    }
+
+    @Test
+    public void Tasks_ShowsNonEmptyMessage() {
+        // Check that the active and completed tasks text is displayed
+        String expectedActiveTaskText = String.format(ApplicationProvider.getApplicationContext()
+                .getString(R.string.statistics_active_tasks), 1);
+        onView(withText(containsString(expectedActiveTaskText))).check(matches(isDisplayed()));
+        String expectedCompletedTaskText = String.format(ApplicationProvider.getApplicationContext()
+                .getString(R.string.statistics_completed_tasks), 1);
+        onView(withText(containsString(expectedCompletedTaskText))).check(matches(isDisplayed()));
     }
 }
