@@ -16,8 +16,8 @@
 
 package com.example.android.architecture.blueprints.todoapp;
 
-import android.databinding.BindingAdapter;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.databinding.BindingAdapter;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.example.android.architecture.blueprints.todoapp.tasks.TasksViewModel;
 
