@@ -17,11 +17,7 @@
 package com.example.android.architecture.blueprints.todoapp.statistics;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
 import android.content.Context;
-import android.databinding.Bindable;
-import android.databinding.ObservableBoolean;
-import android.databinding.ObservableField;
 
 import com.example.android.architecture.blueprints.todoapp.R;
 import com.example.android.architecture.blueprints.todoapp.data.Task;
@@ -29,6 +25,13 @@ import com.example.android.architecture.blueprints.todoapp.data.source.TasksData
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository;
 
 import java.util.List;
+
+import androidx.databinding.Bindable;
+import androidx.databinding.ObservableBoolean;
+import androidx.databinding.ObservableField;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
 
 /**
  * Exposes the data to be used in the statistics screen.
@@ -40,18 +43,15 @@ import java.util.List;
  */
 public class StatisticsViewModel extends AndroidViewModel {
 
-    public final ObservableBoolean dataLoading = new ObservableBoolean(false);
+    private final MutableLiveData<Boolean> mDataLoading = new MutableLiveData<>();
 
-    public final ObservableBoolean error = new ObservableBoolean(false);
+    private final MutableLiveData<Boolean> mError = new MutableLiveData<>();
 
-    public final ObservableField<String> numberOfActiveTasks = new ObservableField<>();
+    private final MutableLiveData<String> mActiveTasks = new MutableLiveData<>();
 
-    public final ObservableField<String> numberOfCompletedTasks = new ObservableField<>();
+    private final MutableLiveData<String> mCompletedTasks = new MutableLiveData<>();
 
-    /**
-     * Controls whether the stats are shown or a "No data" message.
-     */
-    public final ObservableBoolean empty = new ObservableBoolean();
+    private final MutableLiveData mEmpty = new MutableLiveData();
 
     private int mNumberOfActiveTasks = 0;
 
@@ -72,23 +72,48 @@ public class StatisticsViewModel extends AndroidViewModel {
     }
 
     public void loadStatistics() {
-        dataLoading.set(true);
+        mDataLoading.setValue(true);
 
         mTasksRepository.getTasks(new TasksDataSource.LoadTasksCallback() {
             @Override
             public void onTasksLoaded(List<Task> tasks) {
-                error.set(false);
+                mError.setValue(false);
                 computeStats(tasks);
             }
 
             @Override
             public void onDataNotAvailable() {
-                error.set(true);
+                mError.setValue(true);
                 mNumberOfActiveTasks = 0;
                 mNumberOfCompletedTasks = 0;
                 updateDataBindingObservables();
             }
         });
+    }
+
+    // LiveData getters
+
+    public LiveData<Boolean> getDataLoading() {
+        return mDataLoading;
+    }
+
+    public LiveData<Boolean> getError() {
+        return mError;
+    }
+
+    public LiveData<String> getNumberOfActiveTasks() {
+        return mActiveTasks;
+    }
+
+    public LiveData<String> getNumberOfCompletedTasks() {
+        return mCompletedTasks;
+    }
+
+    /**
+     * Controls whether the stats are shown or a "No data" message.
+     */
+    public LiveData<Boolean> getEmpty() {
+        return mEmpty;
     }
 
     /**
@@ -112,12 +137,12 @@ public class StatisticsViewModel extends AndroidViewModel {
     }
 
     private void updateDataBindingObservables() {
-        numberOfCompletedTasks.set(
+        mCompletedTasks.setValue(
                 mContext.getString(R.string.statistics_completed_tasks, mNumberOfCompletedTasks));
-        numberOfActiveTasks.set(
+        mActiveTasks.setValue(
                 mContext.getString(R.string.statistics_active_tasks, mNumberOfActiveTasks));
-        empty.set(mNumberOfActiveTasks + mNumberOfCompletedTasks == 0);
-        dataLoading.set(false);
+        mEmpty.setValue(mNumberOfActiveTasks + mNumberOfCompletedTasks == 0);
+        mDataLoading.setValue(false);
 
     }
 }
