@@ -16,9 +16,9 @@
 
 package com.example.android.architecture.blueprints.todoapp;
 
-import android.arch.lifecycle.Lifecycle;
-import android.arch.lifecycle.LifecycleOwner;
-import android.arch.lifecycle.LifecycleRegistry;
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.LifecycleOwner;
+import androidx.lifecycle.LifecycleRegistry;
 
 
 public class TestUtils {
@@ -40,4 +40,6 @@ public class TestUtils {
             return mRegistry;
         }
     };
+
+
 }
