@@ -17,16 +17,28 @@
 package com.example.android.architecture.blueprints.todoapp.taskdetail;
 
 
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertFalse;
+
+import static org.hamcrest.CoreMatchers.nullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
 import android.app.Application;
-import android.arch.core.executor.testing.InstantTaskExecutorRule;
 import android.content.res.Resources;
 
+import com.example.android.architecture.blueprints.todoapp.Event;
+import com.example.android.architecture.blueprints.todoapp.LiveDataTestUtil;
 import com.example.android.architecture.blueprints.todoapp.R;
-import com.example.android.architecture.blueprints.todoapp.SnackbarMessage;
 import com.example.android.architecture.blueprints.todoapp.data.Task;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource;
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository;
 
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -35,15 +47,7 @@ import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertFalse;
-import static org.hamcrest.CoreMatchers.nullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
 
 /**
  * Unit tests for the implementation of {@link TaskDetailViewModel}
@@ -107,8 +111,8 @@ public class TaskDetailViewModelTest {
         setupViewModelRepositoryCallback();
 
         // Then verify that the view was notified
-        assertEquals(mTaskDetailViewModel.task.get().getTitle(), mTask.getTitle());
-        assertEquals(mTaskDetailViewModel.task.get().getDescription(), mTask.getDescription());
+        assertEquals(mTaskDetailViewModel.getTask().getValue().getTitle(), mTask.getTitle());
+        assertEquals(mTaskDetailViewModel.getTask().getValue().getDescription(), mTask.getDescription());
     }
 
     @Test
@@ -123,33 +127,37 @@ public class TaskDetailViewModelTest {
     }
 
     @Test
-    public void completeTask() {
+    public void completeTask() throws InterruptedException {
         setupViewModelRepositoryCallback();
 
         // When the ViewModel is asked to complete the task
         mTaskDetailViewModel.setCompleted(true);
 
         // Then a request is sent to the task repository and the UI is updated
-        verify(mTasksRepository).completeTask(mTask);
-        assertThat(mTaskDetailViewModel.getSnackbarMessage().getValue(),
-                is(R.string.task_marked_complete));
+        Event<Integer> value = LiveDataTestUtil.getValue(mTaskDetailViewModel.getSnackbarMessage());
+        Assert.assertEquals(
+                (long) value.getContentIfNotHandled(),
+                R.string.task_marked_complete
+        );
     }
 
     @Test
-    public void activateTask() {
+    public void activateTask() throws InterruptedException {
         setupViewModelRepositoryCallback();
 
         // When the ViewModel is asked to complete the task
         mTaskDetailViewModel.setCompleted(false);
 
         // Then a request is sent to the task repository and the UI is updated
-        verify(mTasksRepository).activateTask(mTask);
-        assertThat(mTaskDetailViewModel.getSnackbarMessage().getValue(),
-                is(R.string.task_marked_active));
+        Event<Integer> value = LiveDataTestUtil.getValue(mTaskDetailViewModel.getSnackbarMessage());
+        Assert.assertEquals(
+                (long) value.getContentIfNotHandled(),
+                R.string.task_marked_active
+        );
     }
 
     @Test
-    public void TaskDetailViewModel_repositoryError() {
+    public void TaskDetailViewModel_repositoryError() throws InterruptedException {
         // Given an initialized ViewModel with an active task
         mViewModelCallback = mock(TasksDataSource.GetTaskCallback.class);
 
@@ -162,21 +170,21 @@ public class TaskDetailViewModelTest {
         mGetTaskCallbackCaptor.getValue().onDataNotAvailable(); // Trigger callback error
 
         // Then verify that data is not available
-        assertFalse(mTaskDetailViewModel.isDataAvailable());
+        assertFalse(LiveDataTestUtil.getValue(mTaskDetailViewModel.getIsDataAvailable()));
     }
 
     @Test
-    public void TaskDetailViewModel_repositoryNull() {
+    public void TaskDetailViewModel_repositoryNull() throws InterruptedException {
         setupViewModelRepositoryCallback();
 
         // When the repository returns a null task
         mGetTaskCallbackCaptor.getValue().onTaskLoaded(null); // Trigger callback error
 
         // Then verify that data is not available
-        assertFalse(mTaskDetailViewModel.isDataAvailable());
+        assertFalse(LiveDataTestUtil.getValue(mTaskDetailViewModel.getIsDataAvailable()));
 
         // Then task detail UI is shown
-        assertThat(mTaskDetailViewModel.task.get(), is(nullValue()));
+        assertThat(mTaskDetailViewModel.getTask().getValue(), is(nullValue()));
     }
 
     private void setupViewModelRepositoryCallback() {
@@ -189,14 +197,5 @@ public class TaskDetailViewModelTest {
         verify(mTasksRepository).getTask(eq(mTask.getId()), mGetTaskCallbackCaptor.capture());
 
         mGetTaskCallbackCaptor.getValue().onTaskLoaded(mTask); // Trigger callback
-    }
-
-    @Test
-    public void updateSnackbar_nullValue() {
-        // Before setting the Snackbar text, get its current value
-        SnackbarMessage snackbarText = mTaskDetailViewModel.getSnackbarMessage();
-
-        // Check that the value is null
-        assertThat("Snackbar text does not match", snackbarText.getValue(), is(nullValue()));
     }
 }
